requirements.txt
django>=4.1.4,<4.2
python-dotenv
mysqlclient>=2.1.1,<2.2
djangorestframework